get_status route: add first iteration of the get status route, that is able to only send the process_id of the current agent
