+++ src/http_server/routes/get_status.rs
@@ -1,24 +1,13 @@
-use std::sync::Arc;
+use std::sync::{Arc};
 use axum::{Extension, Json};
-use serde::Serialize;
-use crate::agent_infos::AgentInfos;
+use crate::agent_infos::AgentData;
 
-#[derive(Serialize, Clone)]
-pub struct User {
-    name: String,
-    username: String,
-}
+pub async fn get_status(state: Extension<Arc<AgentData>>) -> Json<AgentData> {
+    let pid = state.get_pid();
 
-pub async fn get_status(state: Extension<AgentInfos>) -> Json<Vec<User>> {
-    let users: Vec<User> = vec![
-        User {
-            name: "Alice".to_string(),
-            username: "alice".to_string(),
-        },
-        User {
-            name: "Bob".to_string(),
-            username: "bob".to_string(),
-        },
-    ];
-    Json(users)
+    Json(
+        AgentData {
+            process_id: pid
+        }
+    )
 }
